test(opportunity): split lifecycle test into standalone spec file
The mega-spec file opportunity.graph.spec.ts has a pre-existing module-
load issue ("describe is not a function" at top-level describe) in local
dev environments — splitting the lifecycle test into its own file keeps
it runnable independently.

- Move the chained approve_introduction → queueNegotiateExisting →
  negotiate_existing → non-introducer notification test to
  introducer-gating-lifecycle.spec.ts
- Revert the mega-spec addition from the prior commit
