+++ packages/protocol/src/opportunity/tests/opportunity.graph.spec.ts
@@ -2396,162 +2396,6 @@ describe('Opportunity Graph', () => {
       expect(negotiateJobsEnqueued[0].opportunityId).toBe('opp-456');
     });
   });
-
-  // ─── Introducer gating lifecycle ─────────────────────────────────────────────
-  describe('introducer gating lifecycle', () => {
-    test('latent opportunity: gate blocks → approve → queue fires → negotiation runs → non-introducers notified', async () => {
-      // Stateful mock DB: tracks actors so an approve mutation is visible to the
-      // subsequent negotiate_existing call (chained via the queueNegotiateExisting callback).
-      const OPP_ID = 'opp-lifecycle';
-      const state = {
-        actors: [
-          {
-            userId: 'patient-user' as Id<'users'>,
-            role: 'patient' as const,
-            networkId: 'idx-1' as Id<'networks'>,
-            intentId: 'intent-patient' as Id<'intents'>,
-            approved: undefined,
-          },
-          {
-            userId: 'agent-user' as Id<'users'>,
-            role: 'agent' as const,
-            networkId: 'idx-1' as Id<'networks'>,
-            intentId: 'intent-agent' as Id<'intents'>,
-            approved: undefined,
-          },
-          {
-            userId: 'introducer-user' as Id<'users'>,
-            role: 'introducer' as const,
-            networkId: 'idx-1' as Id<'networks'>,
-            intentId: undefined,
-            approved: false,
-          },
-        ] as OpportunityActor[],
-      };
-      const currentOpp = (): Opportunity => ({
-        id: OPP_ID,
-        detection: { source: 'manual' as const, createdBy: 'introducer-user' as Id<'users'> },
-        actors: state.actors,
-        interpretation: { reasoning: 'Great match.' },
-        context: null,
-        confidence: 0.9,
-        status: 'latent' as const,
-        createdAt: new Date(),
-        updatedAt: new Date(),
-        expiresAt: null,
-      }) as unknown as Opportunity;
-
-      const negotiationInvocations: unknown[] = [];
-      const notifiedUserIds: string[] = [];
-
-      // Returns acceptance so the non-introducer notification branch fires.
-      const mockNegotiationGraph = {
-        invoke: async (input: unknown) => {
-          negotiationInvocations.push(input);
-          return {
-            outcome: {
-              hasOpportunity: true,
-              agreedRoles: [
-                { userId: 'patient-user', role: 'patient' as const },
-                { userId: 'agent-user', role: 'agent' as const },
-              ],
-              reasoning: 'Accepted.',
-              turnCount: 2,
-            },
-          };
-        },
-      };
-
-      const mockDb: OpportunityGraphDatabase = {
-        getProfile: () => Promise.resolve(null),
-        createOpportunity: (data) => Promise.resolve({ id: 'unused', ...data, status: data.status ?? 'latent', createdAt: new Date(), updatedAt: new Date(), expiresAt: null }),
-        opportunityExistsBetweenActors: () => Promise.resolve(false),
-        getAcceptedOpportunitiesBetweenActors: () => Promise.resolve([]),
-        getOpportunityBetweenActors: () => Promise.resolve(null),
-        findOverlappingOpportunities: () => Promise.resolve([]),
-        getUserIndexIds: () => Promise.resolve(['idx-1'] as Id<'networks'>[]),
-        getNetworkMemberships: async () => [],
-        getActiveIntents: (userId: string) => Promise.resolve([
-          { id: `intent-${userId}` as Id<'intents'>, payload: `Intent for ${userId}`, summary: null, createdAt: new Date() },
-        ]),
-        getNetworkIdsForIntent: () => Promise.resolve([]),
-        getNetwork: () => Promise.resolve({ id: 'idx-1', title: 'Test Index' }),
-        getNetworkMemberCount: () => Promise.resolve(2),
-        getIntentIndexScores: async () => [],
-        getNetworkMemberContext: async () => null,
-        getUser: (id: string) => Promise.resolve({ id, name: `User ${id}`, email: `${id}@example.com` }),
-        isNetworkMember: () => Promise.resolve(false),
-        isIndexOwner: () => Promise.resolve(false),
-        getOpportunity: (id: string) => id === OPP_ID ? Promise.resolve(currentOpp()) : Promise.resolve(null),
-        getOpportunitiesForUser: () => Promise.resolve([]),
-        updateOpportunityStatus: () => Promise.resolve(null),
-        updateOpportunityActorApproval: (_id, userId, approved) => {
-          state.actors = state.actors.map((a) =>
-            a.role === 'introducer' && a.userId === userId ? { ...a, approved } : a,
-          );
-          return Promise.resolve(currentOpp());
-        },
-        getIntent: () => Promise.resolve(null),
-      };
-
-      const mockEmbedder = { generate: async () => dummyEmbedding } as unknown as Embedder;
-      const mockHyde = { invoke: async () => ({ hydeEmbeddings: {} }) };
-      const queueNotification = async (_oppId: string, userId: string) => {
-        notifiedUserIds.push(userId);
-      };
-
-      let compiledGraph: ReturnType<OpportunityGraphFactory['createGraph']>;
-      const queueNegotiateExistingCalls: Array<{ opportunityId: string; userId: string }> = [];
-      // Simulate BullMQ: the callback triggers negotiate_existing on the same compiled graph,
-      // exactly like the backend queue worker (handleNegotiateExisting).
-      const queueNegotiateExisting = async (opportunityId: string, userId: string) => {
-        queueNegotiateExistingCalls.push({ opportunityId, userId });
-        await compiledGraph.invoke({
-          userId: userId as Id<'users'>,
-          operationMode: 'negotiate_existing' as const,
-          opportunityId,
-          options: {},
-        });
-      };
-
-      const factory = new OpportunityGraphFactory(
-        mockDb,
-        mockEmbedder,
-        mockHyde,
-        createMockEvaluator(),
-        queueNotification,
-        mockNegotiationGraph,
-        undefined, // agentDispatcher
-        queueNegotiateExisting,
-      );
-      compiledGraph = factory.createGraph();
-
-      // ─── Phase 1: gate is armed (approved: false). No approve_introduction yet. ────
-      // Note: the gate test at line 1912 already exercises the negotiate-node skip.
-      // Here we assert only that no negotiation has run yet before approval.
-      expect(negotiationInvocations).toHaveLength(0);
-
-      // ─── Phase 2: introducer approves → callback should fire ──────────────────────
-      await compiledGraph.invoke({
-        userId: 'introducer-user' as Id<'users'>,
-        opportunityId: OPP_ID,
-        operationMode: 'approve_introduction' as const,
-      });
-
-      // Approval was persisted in state.
-      const postApprovalIntroducer = state.actors.find((a) => a.role === 'introducer');
-      expect(postApprovalIntroducer?.approved).toBe(true);
-
-      // Callback fired with the correct payload.
-      expect(queueNegotiateExistingCalls).toEqual([{ opportunityId: OPP_ID, userId: 'introducer-user' }]);
-
-      // ─── Phase 3: callback ran negotiate_existing → negotiation + notifications ───
-      expect(negotiationInvocations.length).toBeGreaterThan(0);
-      expect(notifiedUserIds).toContain('patient-user');
-      expect(notifiedUserIds).toContain('agent-user');
-      expect(notifiedUserIds).not.toContain('introducer-user');
-    });
-  });
 });
 
 // ─── buildDiscovererContext tests ───────────────────────────────────────────
